Add shell scripts to start and stop Docker containers
Created scripts to start and stop Docker containers for the applipress application. The 'start_ap.sh' and 'stop_ap.sh' scripts were introduced to provide a simplified way of controlling Docker containers. The scripts determine the environment (Linux, macOS, or other) and then set appropriate user IDs and groupIDs. Scripts are also equipped with checks to test the availability of docker compose or docker-compose tools and to provide user-friendly output in case these tools are not installed. The .gitignore file to exclude IDE-generated files was also updated. The Readme file has been updated to reflect this operational change. This approach provides a more consistent user experience across different platforms and simplifies the process of starting and stopping the application.